EOSM: mixup Makefiles to get ready for EOS M arrival: - added EOS M to mvr.h, consts.h, and gui.h - fixed header of stubs file (no stubs verified yet, just taken from 5d3 for now)
In EOS M installer Makefile, I changed the names of some targets (listed below) to fix a duplicate target error:
- make all       ==>   make all_eosm
- make clean     ==>   make_clean_eosm
- make install   ==>   make install_eosm
This is an issue because there is already a target for all, clean, install, etc in the main Makefile, but we have defined new versions of them in the installer Makefile.

Added EOS M to target prepare_zip in Makefile.inc. It works but this code needs updated since INSTALL.pdf and UserGuide.pdf are missing, plugins folder is empty, and there is no 'cam' folder inside the 'doc' folder.

In main Makefile, I added EOS M to main list of cameras able to be built. Also added required entries for building with 'make fir' and 'make install_fir'. Finally, I added the EOS M to the 'clean' target list.

In reboot-ins.c I fixed it so I'm representing memory addresses with unsigned instead of int (signed).