use RST2LATEX for selecting "rst2latex" program
E.g. on my Debian 7.1 and Ubuntu 13.04 the program
for converting from rst to latex is named "rst2latex".
But the ML build scripts suppose that this program
has name "rst2latex.py".

There is no way easy redefine this name in one point,
so this commit introduces the RST2LATEX makefile variable.